Lint and codestyle
