+++ frontend/form-builder/src/components/Toolbar/ToolBarItems/ToolbarItems.module.css
@@ -1,18 +1,18 @@
 .toolbarItems {
-    display: flex;
-    flex-direction: column;
-    gap: var(--fds-spacing-4);
+  display: flex;
+  flex-direction: column;
+  gap: var(--fds-spacing-4);
 }
 
 @media (max-width: 768px) {
-    .toolbarItems {
-        flex-direction: row;
-        width: 100%;
-        gap: 0;
-        padding: 0;
-    }
+  .toolbarItems {
+    flex-direction: row;
+    width: 100%;
+    gap: 0;
+    padding: 0;
+  }
 
-    .toolbarHeading {
-        display: none;
-    }
+  .toolbarHeading {
+    display: none;
+  }
 }
