+++ frontend/main-app/src/components/UserDropdown/UserDropdown.module.css
@@ -1,5 +1,5 @@
 .dropdownButton {
-  border: 1px solid var(--fds-semantic-border-neutral-default);;
+  border: 1px solid var(--fds-semantic-border-neutral-default);
 }
 
 .dropdownMenu {
